Ex10: Refactored app.py, placed @requires_role in external file
